+++ src/hook/generateThemes.ts
@@ -52,8 +52,6 @@ const utils = new Utils();
 const configuration: Configuration = {
   darkContrast: "medium",
   lightContrast: "medium",
-  darkWorkbench: "material",
-  lightWorkbench: "material",
   highContrast: false,
 };